+++ CHANGELOG.md
@@ -18,6 +18,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 parameter. 
 - `CircuitBuilder` accepts multiple (qu)bit registers through `add_register` method.
 - Add `interaction_graph` as a property of the `Circuit`.
+- Add `mapping` as an attribute of the `Circuit`, which contains an identity mapping, prior to any 
+arbitrarily applied mapper pass.
 
 ## [ 0.9.0 ] - [ 2025-12-19 ]